Apysaics and Talataerim adopt Clan
Tweak to the adopt Feudal/Clan decision, so that Apysaics and Talataerim become Clan unless they already have a Feudal liege. I hope this ends up as a change in the patch notes! Now back to the history mines...

(Will have to be updated every time a new Clan-based heritage is added, and if pdx makes another playable Feudal/Clan-tier or Tribal-tier government)