"King in the Golden Hall" Event Chain rarer
The event chain previously had a 50% chance of occurring every year.
Reduced chance to 5% (Thus it still has an 80% chance to occur within 30 years), and added a 15 year grace period after gamestart so it does not interfere with story content.

In the future, the chain should be looked at and improved, especially regarding its localization