Minor and Major bugfixes
The minor bugfixes are just that, minor, but there were two major ones that involved making some decisions that may require some review.

For 'The Succession of Pezarsan' I tried my best to stick with what I thought was the intent based on the loc. This meant adding Zakakha as a tributary and partner in the war while updating what the actual characters in charge of the allies names are.

For Uri VI lacking his VI, I just set his name as Uri, #- out the event which did so.