Succession of Pezaran - Add Clarity
Speaking of clarity, despite having the events all around them translated, this one is in English in both the French and Russian files. IDK why, but that's why I updated their files too, so that when the translators get around to it, they are working from the right, more clear, correct character referencing version.