Kârna/Jelut History Work for Angmar Start
Adds a large batch of history work for Mid TA Kârna (future Narak). Implements a gondorian tributary kingdom, as well as setting the ground for future work on the Jelut Invasion.

Committing this now so I can focus on bugfixing etc from now on, will pick it back up in a while