Major Tweaks to Trans-Haradric Merchants, Trade and Legacies
Major tweaks to the trans-haradric merchants travel option and legacy.

The trait should now be much easier to get, and the legacies have different effects to further improve it.

You can no longer unlock the option through a stewardship perk, but it is instead unlocked through the second haradric legacy regardless of innovation.