Changes to Noldor shield + gondor sword + dwarf helmet 03
I've slightly changed the noldor shield model and matched the colour of the golden variation to that of the new armour. Also slightly modified the shape of the dwarven helmet 03 so that Balin's ears won't clip through. Finally, slightly modified the hilt of the gondorian sword to better represent the movie prop.