Prevent desert event option from triggering everywhere
Base game trigger checks for being in the Sahara desert but not at an oasis. If so, the event only has one option which is negative. By commenting out the Sahara part, the mod's version only checks that you're not at an oasis, and the event has only the negative option everywhere else. Using the Dune Sea region seems like a good alternative, but feel free to refine further if necessary.